[Feature] Add basic SPARSEMEM definition for later
This is a game about data structures and algorithms.

In order to use SPARSEMEM instead of FLATMEM and DISCONTMEM, Sandix need some
data structures (struct sparsemem_section, etc.) to save all info about how
to use mem_map and phys memory.

A lot of others macros are not added in this commit, since I do know what macros
we need. I need to revisit this part after FLATMEM and DISCONTMEM.