[Bug] Fix movs bug in 16-bit bootloader, add 0x67
The movsl instruction moves [%ds:%(e)si] to [%es:%(e)di].
Here, we have two versions:
	1) [%ds:%si]  --> [%es:%di]
	2) [%ds:%esi] --> [%es:%edi]

The 1) version can address up to 0xffff:0xffff, which is 0x10ffef, which
is 64KB. So this is not enough to load kernel image.

The 2) version works if and only if the movs has 0x66 and 0x67 prefixes.
And we can set %ds=%es=0, then %esi and %edi can address up to 4GB. This
is very COOL.